# frontend/src/api-client/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "YAESandBox API (Public)",
    "description": "供前端和外部使用的公开 API 文档。",
    "version": "v1"
  },
  "paths": {
...
    "/api/Persistence/save": {
      "post": {
        "tags": [
          "Persistence"
        ],
        "summary": "保存当前 YAESandBox 的完整状态（包括所有 Block、WorldState、GameState）到一个 JSON 文件。\r\n客户端可以（可选地）在请求体中提供需要一同保存的“盲存”数据。",
        "requestBody": {
          "description": "（可选）客户端提供的任意 JSON 格式的盲存数据，将原样保存在存档文件中。",
          "content": {
            "application/json": {
              "schema": { }
            },
            "text/json": {
              "schema": { }
            },
            "application/*+json": {
              "schema": { }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功生成并返回存档文件。",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "保存状态时发生内部服务器错误。"
          }
        }
      }
    },
    "/api/Persistence/load": {
      "post": {
        "tags": [
          "Persistence"
        ],
        "summary": "从上传的 JSON 存档文件加载 YAESandBox 的状态。\r\n这将完全替换当前内存中的所有 Block、WorldState 和 GameState。\r\n成功加载后，将返回存档文件中包含的“盲存”数据（如果存在）。",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "archiveFile": {
                    "type": "string",
                    "description": "包含 YAESandBox 存档的 JSON 文件。",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "archiveFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功加载状态，并返回盲存数据。",
            "content": {
              "text/plain": {
                "schema": { }
              },
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              }
            }
          },
          "400": {
            "description": "没有上传文件，或者上传的文件格式无效 (非 JSON 或内容损坏)。",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "加载状态时发生内部服务器错误。"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}